fix: preserve observation dtype and return chronological transitions
- Remove forced np.uint8 cast in collect_actor_transitions and
  collect_random_transitions so float observations are not quantized.
- ReplayBuffer._allocate() now accepts obs_dtype parameter; add()
  forwards the actual observation dtype.
- load_state_dict() passes stored obs dtype through both tensorized
  and legacy restore paths.
- transitions property returns chronological order after wrap-around
  instead of physical ring-buffer order.
- Update _obs/_next_obs type annotations to generic NDArray.

+++ src/tiny_dreamer_highway/data/collect_random_rollouts.py
@@ -64,10 +64,10 @@ def collect_random_transitions(
             done = bool(terminated or truncated)
             replay_buffer.add(
                 Transition(
-                    observation=np.asarray(observation, dtype=np.uint8),
+                    observation=np.asarray(observation),
                     action=stored_action,
                     reward=float(reward),
-                    next_observation=np.asarray(next_observation, dtype=np.uint8),
+                    next_observation=np.asarray(next_observation),
                     done=done,
                     terminated=bool(terminated),
                     truncated=bool(truncated),
